`omero-server-config-update.sh` replaced by `omero load --glob /opt/omero/server/config/*.omero`
This was broken when the image was switched to Python 3- the Ansible role was switched to use the new built-in `load --glob` feature in OMERO.py but this Docker image wasn't updated to match.

This means the current released 5.6 omero-server Docker image does not load /opt/omero/server/config/*.omero.